I'd say it's enough.

+++ src/main/java/net/kodehawa/mantarobot/commands/rpg/CurrencyManager.java
@@ -8,7 +8,7 @@ public class CurrencyManager {
 	public static final double USD_ESTIMATIVE = 1500000000000D; //1.5TRI
 
 	public static double creditsWorth() {
-		return (double) USD_ESTIMATIVE / MantaroData.getData().get().users.values().stream()
+		return USD_ESTIMATIVE / MantaroData.getData().get().users.values().stream()
 			.mapToDouble(EntityPlayer::getMoney).sum();
 	}
 }
